+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/AdditionalFieldServiceImplTest.java
@@ -6,6 +6,7 @@ import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.issue.fields.layout.field.FieldLayout;
 import com.atlassian.jira.issue.fields.layout.field.FieldLayoutItem;
 import com.atlassian.jira.issue.fields.layout.field.FieldLayoutManager;
+import de.codescape.jira.plugins.scrumpoker.model.AdditionalField;
 import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
 import org.junit.Rule;
 import org.junit.Test;
@@ -16,19 +17,18 @@ import org.mockito.junit.MockitoRule;
 import webwork.action.Action;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.*;
 
 public class AdditionalFieldServiceImplTest {
 
     private static final String MISSING_FIELD = "MISSING_FIELD";
     private static final String EXISTING_FIELD = "EXISTING_FIELD";
-    private static final String NEW_FIELD = "NEW_FIELD";
 
     @Rule
     public MockitoRule rule = MockitoJUnit.rule();
@@ -63,21 +63,51 @@ public class AdditionalFieldServiceImplTest {
     @Mock
     private FieldLayoutItem fieldLayoutItem;
 
-    // tests for supportedCustomFields()
+    /* tests for supportedCustomFields() */
 
     @Test
-    public void supportedCustomFieldsDelegatesToCustomFieldManager() {
-        List<CustomField> expectedFields = new ArrayList<>();
-        expectedFields.add(customField);
-        when(customFieldManager.getCustomFieldObjects()).thenReturn(expectedFields);
+    public void supportedCustomFieldsPreservesOrderOfConfiguredCustomFields() {
+        when(globalSettingsService.load()).thenReturn(globalSettings);
+        when(globalSettings.getAdditionalFields()).thenReturn("Third Field,Second Field");
+        CustomField field1 = mockCustomField("First Field");
+        CustomField field2 = mockCustomField("Second Field");
+        CustomField field3 = mockCustomField("Third Field");
+        CustomField field4 = mockCustomField("Fourth Field");
+        ArrayList<CustomField> customFields = new ArrayList<>(Arrays.asList(field1, field2, field3, field4));
+        when(customFieldManager.getCustomFieldObjects()).thenReturn(customFields);
 
-        List<CustomField> fields = service.supportedCustomFields();
+        List<AdditionalField> fields = service.supportedCustomFields();
 
-        assertThat(fields.size(), is(1));
-        assertThat(fields, contains(customField));
+        assertThat(fields.size(), is(4));
+        assertThat(fields, contains(
+            new AdditionalField(field3, true),
+            new AdditionalField(field2, true),
+            new AdditionalField(field1, false),
+            new AdditionalField(field4, false)));
     }
 
-    // tests for configuredCustomFields()
+    @Test
+    public void supportedCustomFieldsPreservesOrderOfExistingCustomFields() {
+        when(globalSettingsService.load()).thenReturn(globalSettings);
+        when(globalSettings.getAdditionalFields()).thenReturn(null);
+        CustomField field1 = mockCustomField("First Field");
+        CustomField field2 = mockCustomField("Second Field");
+        CustomField field3 = mockCustomField("Third Field");
+        CustomField field4 = mockCustomField("Fourth Field");
+        ArrayList<CustomField> customFields = new ArrayList<>(Arrays.asList(field1, field2, field3, field4));
+        when(customFieldManager.getCustomFieldObjects()).thenReturn(customFields);
+
+        List<AdditionalField> fields = service.supportedCustomFields();
+
+        assertThat(fields.size(), is(4));
+        assertThat(fields, contains(
+            new AdditionalField(field1, false),
+            new AdditionalField(field2, false),
+            new AdditionalField(field3, false),
+            new AdditionalField(field4, false)));
+    }
+
+    /* tests for configuredCustomFields() */
 
     @Test
     public void configuredCustomFieldsIgnoresMissingFields() {
@@ -101,81 +131,7 @@ public class AdditionalFieldServiceImplTest {
         assertThat(fields, contains(customField));
     }
 
-    // tests for addConfiguredField())
-
-    @Test
-    public void addConfiguredFieldIgnoresEmptyString() {
-        service.addConfiguredField("");
-        verify(globalSettingsService, never()).persist(any());
-    }
-
-    @Test
-    public void addConfiguredFieldIgnoresNullValue() {
-        service.addConfiguredField(null);
-        verify(globalSettingsService, never()).persist(any());
-    }
-
-    @Test
-    public void addConfiguredFieldSupportsInitialEmptyList() {
-        when(globalSettingsService.load()).thenReturn(globalSettings);
-        when(globalSettings.getAdditionalFields()).thenReturn(null);
-
-        service.addConfiguredField(NEW_FIELD);
-
-        verify(globalSettings).setAdditionalFields(NEW_FIELD);
-        verify(globalSettingsService).persist(globalSettings);
-    }
-
-    @Test
-    public void addConfiguredFieldIgnoresRedundantValues() {
-        when(globalSettingsService.load()).thenReturn(globalSettings);
-        when(globalSettings.getAdditionalFields()).thenReturn(EXISTING_FIELD);
-
-        service.addConfiguredField(EXISTING_FIELD);
-
-        verify(globalSettings).setAdditionalFields(EXISTING_FIELD);
-        verify(globalSettingsService).persist(globalSettings);
-    }
-
-    @Test
-    public void addConfiguredFieldAppendsNewValues() {
-        when(globalSettingsService.load()).thenReturn(globalSettings);
-        when(globalSettings.getAdditionalFields()).thenReturn(EXISTING_FIELD);
-        when(customFieldManager.getCustomFieldObject(EXISTING_FIELD)).thenReturn(customField);
-
-        service.addConfiguredField(NEW_FIELD);
-
-        verify(globalSettings).setAdditionalFields(EXISTING_FIELD + "," + NEW_FIELD);
-        verify(globalSettingsService).persist(globalSettings);
-    }
-
-    // tests for removeConfiguredField()
-
-    @Test
-    public void removeConfiguredFieldRemovesExistingValue() {
-        when(globalSettingsService.load()).thenReturn(globalSettings);
-        when(globalSettings.getAdditionalFields()).thenReturn(EXISTING_FIELD);
-        when(customFieldManager.getCustomFieldObject(EXISTING_FIELD)).thenReturn(customField);
-
-        service.removeConfiguredField(EXISTING_FIELD);
-
-        verify(globalSettings).setAdditionalFields("");
-        verify(globalSettingsService).persist(globalSettings);
-    }
-
-    @Test
-    public void removeConfiguredFieldAlsoRemovesInvalidValue() {
-        when(globalSettingsService.load()).thenReturn(globalSettings);
-        when(globalSettings.getAdditionalFields()).thenReturn(EXISTING_FIELD + "," + MISSING_FIELD);
-        when(customFieldManager.getCustomFieldObject(EXISTING_FIELD)).thenReturn(customField);
-
-        service.removeConfiguredField(EXISTING_FIELD);
-
-        verify(globalSettings).setAdditionalFields("");
-        verify(globalSettingsService).persist(globalSettings);
-    }
-
-    // tests for renderFieldValue()
+    /* tests for renderFieldValue() */
 
     @Test
     public void renderFieldValue() {
@@ -189,6 +145,14 @@ public class AdditionalFieldServiceImplTest {
         verify(fieldLayoutManager, times(1)).getFieldLayout(issue);
         verify(fieldLayout, times(1)).getFieldLayoutItem(customField);
         verify(customField, times(1)).getViewHtml(fieldLayoutItem, action, issue);
+    }
+
+    /* supporting methods */
+
+    private CustomField mockCustomField(String customFieldId) {
+        CustomField customField = mock(CustomField.class);
+        when(customField.getId()).thenReturn(customFieldId);
+        return customField;
     }
 
 }