+++ docs/changelog.md
@@ -17,6 +17,7 @@ category: Administration
 * feat: add internationalization for additional fields on session view (#21)
 * feat: ensure correct displaying with no fields configured (#21)
 * test: add unit test coverage for AdditionalFieldService (#21)
+* feat: add internationalization for additional fields configuration (#21)
 
 ### [20.08.0] - 2020-08-27